rust: use usize for lengths instead of i32
This offers better interop with other Rust libraries.  Rust is safer
for avoiding overflow/unsigned bugs.

+++ rust/src/lib.rs
@@ -133,7 +133,7 @@ impl Ceed {
     /// let ceed = ceed::Ceed::init("/cpu/self/ref/serial");
     /// let vec = ceed.vector(10);
     /// ```
-    pub fn vector(&self, n: i32) -> crate::vector::Vector {
+    pub fn vector(&self, n: usize) -> crate::vector::Vector {
         crate::vector::Vector::create(self, n)
     }
